backend/requirements.txt
fastapi==0.128.8
uvicorn==0.40.0
groq==1.0.0
python-dotenv==1.0.1
pydantic==2.12.5
cloudinary==1.41.0
manim==0.18.1
